Module aws_sdk_serverlessapplicationrepository::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

not_found_exception

See NotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

BadRequestException

One of the parameters in the request is invalid.

ConflictException

The resource already exists.

CreateApplicationError

CreateApplicationVersionError

CreateCloudFormationChangeSetError

CreateCloudFormationTemplateError

DeleteApplicationError

ForbiddenException

The client is not authenticated.

GetApplicationError

GetApplicationPolicyError

GetCloudFormationTemplateError

InternalServerErrorException

The AWS Serverless Application Repository service encountered an internal error.

ListApplicationDependenciesError

ListApplicationVersionsError

ListApplicationsError

NotFoundException

The resource (for example, an access policy statement) specified in the request doesn't exist.

PutApplicationPolicyError

TooManyRequestsException

The client is sending more than the allowed number of requests per unit of time.

UnshareApplicationError

UpdateApplicationError

Enums

CreateApplicationErrorKind

CreateApplicationVersionErrorKind

CreateCloudFormationChangeSetErrorKind

CreateCloudFormationTemplateErrorKind

DeleteApplicationErrorKind

GetApplicationErrorKind

GetApplicationPolicyErrorKind

GetCloudFormationTemplateErrorKind

ListApplicationDependenciesErrorKind

ListApplicationVersionsErrorKind

ListApplicationsErrorKind

PutApplicationPolicyErrorKind

UnshareApplicationErrorKind

UpdateApplicationErrorKind